+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v12.0.10
 * Fix #2378 - cast.lua compound unit tokens are not allowed for this call
+* Fix #2398 - Auras.lua:303: attempt to index a nil value
 
 ## v12.0.9
 * Fix #2371 - Corruption treated as Immolate